eda/requirements-eda.txt
# EDA-only optional dependencies
pandas>=2.1
matplotlib>=3.8
seaborn>=0.13
marimo>=0.22
sqlalchemy>=2.0
PyYAML>=6.0
psycopg[binary]>=3.1
